requirements.txt
annotated-types==0.6.0
anyio==4.3.0
bcrypt==4.1.2
beanie==1.25.0
click==8.1.7
dnspython==2.6.1
email_validator==2.1.1
fastapi==0.110.0
h11==0.14.0
httptools==0.6.1
idna==3.6
iniconfig==2.0.0
Jinja2==3.1.3
lazy-model==0.2.0
MarkupSafe==2.1.5
motor==3.3.2
packaging==24.0
pluggy==1.4.0
pydantic==2.6.4
pydantic_core==2.16.3
PyJWT==2.8.0
pymongo==4.6.2
pytest==8.1.1
pytest-asyncio==0.23.5.post1
python-decouple==3.8
python-dotenv==1.0.1
PyYAML==6.0.1
sniffio==1.3.1
starlette==0.36.3
toml==0.10.2
typing_extensions==4.10.0
uvicorn==0.28.0
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0

torch~=2.2.1
